+++ CHANGELOG.md
@@ -28,6 +28,7 @@ New Behavior: InfluxDB 1.x returned data as an epoch float value (e.g. 166722857
  - Add ability to select Outputs in Output Modules
  - Add ability to average measurements in Verification Function
  - Add ability to select Output Measurements in Equation Functions
+ - Add Output: Sequent Microsystems 8-Relay HAT for Raspberry Pi
 
 ### Bugfixes